pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "darkreconx"
version = "0.1.0"
description = "Modular OSINT reconnaissance framework with Tor support"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "DarkReconX Contributors"}
]
dependencies = [
    "requests>=2.31.0",
    "PySocks>=1.7.1",
    "stem>=1.8.2",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyfiglet>=1.0.2",
    "pyyaml>=6.0.1",
    "python-whois>=0.9.4",
    "dnspython>=2.6.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[tool.black]
line-length = 127
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | venv
)/
'''

[tool.isort]
profile = "black"
line_length = 127
known_first_party = ["core", "config", "modules", "cli", "api", "tests"]
skip_glob = ["*/migrations/*"]
include_trailing_comma = true
use_parentheses = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_unimported = false
ignore_missing_imports = true
exclude = [".venv", "venv", "build", "dist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
pythonpath = ["."]
markers = [
    "integration: marks tests as integration (networked)",
    "asyncio: marks tests as async",
    "slow: marks tests as slow",
]

[tool.coverage.run]
source = ["core", "config", "modules", "cli"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.setuptools]
packages = ["core", "config", "modules", "cli", "api"]
